+++ src/nd2/_pysdk/_decode.py
